Extract shared types into common library
