+++ .vscode/launch.json
@@ -8,7 +8,7 @@
             "type": "lldb",
             "request": "launch",
             "name": "Debug executable 'xicm-portal-gtk'",
-            "program": "${workspaceFolder}/build/xyz-iinuwa-credential-manager-portal-gtk/src/xyz-iinuwa-credential-manager-portal-gtk",
+            "program": "${workspaceFolder}/build/xyz-iinuwa-credential-manager-portal-gtk/debug/xyz-iinuwa-credential-manager-portal-gtk",
             "args": [],
             "env": {
                 "GSETTINGS_SCHEMA_DIR": "${workspaceFolder}/build/xyz-iinuwa-credential-manager-portal-gtk/data",
