+++ creds-lib/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "creds-lib"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+futures-lite = "2.6.0"
+libwebauthn = "0.2"
+serde = { version = "1", features = ["derive"] }
+zbus = "5.9.0"
