+++ creds-ui/Cargo.toml
@@ -1,12 +1,15 @@
 [package]
-name = "demo-ui"
+name = "creds-ui"
 version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-async-std = "1.13.1"
+async-std = { version = "1.13.1", features = ["unstable"] }
+creds-lib = { path = "../creds-lib" }
+futures-lite = "2.6.0"
 gettext-rs = { version = "0.7", features = ["gettext-system"] }
 gtk = { version = "0.9.6", package = "gtk4", features = ["v4_6"] }
 qrcode = "0.14.1"
 serde = { version = "1.0.219", features = ["derive"] }
 tracing = "0.1.41"
+zbus = "5.9.0"
