+++ creds-ui/src/gui/mod.rs
@@ -1,40 +1,36 @@
 pub mod view_model;
 
-use std::sync::Arc;
 use std::thread;
+use std::{sync::Arc, thread::JoinHandle};
 
 use async_std::{channel::Receiver, sync::Mutex as AsyncMutex};
-use tokio::sync::oneshot;
 
-use crate::credential_service::CredentialServiceClient;
-use crate::model::{Operation, ViewRequest, ViewUpdate};
+use creds_lib::server::ViewRequest;
+use creds_lib::{
+    client::CredentialServiceClient,
+    model::{Operation, ViewUpdate},
+};
 
 use view_model::ViewEvent;
 
 pub(super) fn start_gui_thread<C: CredentialServiceClient + Send + Sync + 'static>(
     rx: Receiver<ViewRequest>,
     client: C,
-) {
-    thread::Builder::new()
-        .name("gui".into())
-        .spawn(move || {
-            let client = Arc::new(AsyncMutex::new(client));
-            // D-Bus received a request and needs a window open
-            while let Ok(view_request) = rx.recv_blocking() {
-                run_gui(client.clone(), view_request);
-            }
-        })
-        .unwrap();
+) -> Result<JoinHandle<()>, std::io::Error> {
+    thread::Builder::new().name("gui".into()).spawn(move || {
+        let client = Arc::new(AsyncMutex::new(client));
+        // D-Bus received a request and needs a window open
+        while let Ok(view_request) = rx.recv_blocking() {
+            run_gui(client.clone(), view_request);
+        }
+    })
 }
 
 fn run_gui<C: CredentialServiceClient + Send + Sync + 'static>(
     client: Arc<AsyncMutex<C>>,
     request: ViewRequest,
 ) {
-    let ViewRequest {
-        operation,
-        signal: response_tx,
-    } = request;
+    let operation = request.operation;
     let (tx_update, rx_update) = async_std::channel::unbounded::<ViewUpdate>();
     let (tx_event, rx_event) = async_std::channel::unbounded::<ViewEvent>();
     let event_loop = async_std::task::spawn(async move {
@@ -46,7 +42,6 @@ fn run_gui<C: CredentialServiceClient + Send + Sync + 'static>(
     view_model::gtk::start_gtk_app(tx_event, rx_update);
 
     async_std::task::block_on(event_loop.cancel());
-    response_tx.send(()).unwrap();
 }
 
 trait GuiClient {
