+++ xyz-iinuwa-credential-manager-portal-gtk/Cargo.toml
@@ -9,6 +9,7 @@ lto = true
 
 [dependencies]
 base64 = "0.22.1"
+creds-lib = { path = "../creds-lib" }
 openssl = "0.10.72"
 ring = "0.17.14"
 serde = { version = "1.0.219", features = ["derive"] }
@@ -17,7 +18,7 @@ serde_json = "1.0.140"
 tracing = "0.1.41"
 tracing-subscriber = "0.3"
 zbus = { version = "5.5.0", default-features = false, features = ["blocking-api", "tokio"] }
-libwebauthn = { git = "https://github.com/linux-credentials/libwebauthn", rev = "34f8a59cb1634175b8baf866e6d30d1869f5a221" }
+libwebauthn = "~0.2.2"
 async-trait = "0.1.88"
 tokio = { version = "1.45.0", features = ["rt-multi-thread"] }
 futures-lite = "2.6.0"
