+++ xyz-iinuwa-credential-manager-portal-gtk/src/main.rs
@@ -2,7 +2,7 @@ mod cbor;
 mod cose;
 mod credential_service;
 mod dbus;
-mod model;
+// mod model;
 mod serde;
 mod webauthn;
